testing if table exists

+++ R/test_identifier.R
@@ -1,9 +1,14 @@
-test_identifier <- function(name, table, variable){
-  tabledata <- read.csv(paste(table, ".csv", sep=""))
-  fun <- function(x) x %in% x[duplicated(x)] 
-  if(length(tabledata[,variable][fun(tabledata[,variable])])!=0){   
-    doppelt<-unique(tabledata[,variable][fun(tabledata[,variable])])
-    out <- tabledata[which(tabledata[,variable]==doppelt),]
-    write.csv(out, paste(name, ".csv", sep=""))
+test_identifier <- function(path_in, name, table, variable){
+  if(!file.exists(paste0(path_in, table, ".csv"))) {
+    message(paste0("table ", table, " does not exist in ", path_in))
   }
-}  
+  else{
+    tabledata <- read.csv(paste(table, ".csv", sep=""))
+    fun <- function(x) x %in% x[duplicated(x)] 
+    if(length(tabledata[,variable][fun(tabledata[,variable])])!=0){   
+      doppelt<-unique(tabledata[,variable][fun(tabledata[,variable])])
+      out <- tabledata[which(tabledata[,variable]==doppelt),]
+      write.csv(out, paste(name, ".csv", sep=""))
+    }
+  }  
+}
